+++ src/Query/BatchQueryResultInterface.php
@@ -33,7 +33,7 @@ use Yiisoft\Db\Exception\InvalidConfigException;
  * @extends Iterator<int, array>
  *
  * @psalm-import-type IndexBy from QueryInterface
- * @psalm-type ResultCallback = Closure(non-empty-list<array>):non-empty-array<array|object>
+ * @psalm-import-type ResultCallback from QueryInterface
  */
 interface BatchQueryResultInterface extends Iterator
 {
@@ -97,6 +97,11 @@ interface BatchQueryResultInterface extends Iterator
      * @param int $value The number of rows to return in each batch.
      */
     public function batchSize(int $value): static;
+
+    /**
+     * @psalm-param IndexBy|null $indexBy
+     */
+    public function indexBy(Closure|string|null $indexBy): static;
 
     /**
      * Sets a callback function to be called for the result of the query.