+++ src/Query/Query.php
@@ -268,8 +268,7 @@ class Query implements QueryInterface
     {
         return $this->db
             ->createBatchQueryResult($this)
-            ->batchSize($batchSize)
-            ->resultCallback($this->index(...));
+            ->batchSize($batchSize);
     }
 
     public function column(): array
@@ -846,12 +845,12 @@ class Query implements QueryInterface
     }
 
     /**
-     * @psalm-param list<array> $rows
+     * @psalm-param array<array> $rows
      *
      * @return array[]|object[]
      *
      * @psalm-return (
-     *     $rows is non-empty-list<array>
+     *     $rows is non-empty-array<array>
      *         ? non-empty-array<array|object>
      *         : array[]|object[]
      * )